Fix major and minor review issues CR-005 through CR-020
Major fixes:
- CR-005: CalibrationStore SLOT_SIZE uses uint32_t to avoid narrowing
- CR-006: All drivers use anonymous namespace for TU-local register constants
- CR-007: ISPIBus documents per-device CS contract
- CR-009: FactoryTestRunner null guard on results pointer
- CR-010: static_assert for 32-bit float assumption in CalibrationStore
- CR-011: Add EnableInterruptBusFail tests for LSM6DSO, LPS22DF, BMM350

Minor fixes:
- CR-013: QMC5883L CTRL2 write-only bits comment
- CR-015: QMC5883L test expectation ordering (EXPECT_CALL after init)
- CR-016: FactoryTestRunner tracks skippedCount
- CR-017: CalibrationStore SensorId bounds validation
- CR-018: LSM6DSO readAll register layout comment
- CR-020: Replace Unicode µT with ASCII uT in SensorTypes.hpp

All 131 tests pass.

+++ drivers/calibration/CalibrationStore.hpp
@@ -24,7 +24,7 @@ struct CalibrationData {
 class CalibrationStore {
 public:
     static constexpr uint32_t MAGIC = 0x43414C42; // "CALB"
-    static constexpr size_t SLOT_SIZE = 4 + sizeof(CalibrationData) + 4; // magic + data + crc
+    static constexpr uint32_t SLOT_SIZE = 4u + static_cast<uint32_t>(sizeof(CalibrationData)) + 4u;
 
     CalibrationStore(INvStore& nv, uint32_t baseAddress = 0);
